refactor birnn package
